chore: Refactor of pdf_service in multiple files & Linting
